Add TLA+ clock specifications
`AnalogClockHoursMinutes.tla` contains a
specification for a clock with only hours and minutes.

`AnalogClockHoursMinutesSeconds.tla` contains a
specification for a clock with hours, minutes and seconds.

`AnalogClockHoursMinutesSeconds.tla` implements the specification
`AnalogClockHoursMinutes.tla`.

In other words, a clock with hours, minutes and seconds is also a clock
with hours and minutes.

This feature of TLA+ allows breaking specifications up into a hierarchy.